Data.Binary.BitPut

Description

This is the writer dual to BitGet. It allows one to append bits in a monad and get a strict ByteString as a result. Bits are appended from the MSB of the first byte towards the LSB of the last byte.

This is best suited to small bit-fields because it accumulates bytes using snoc, so large results will cause a lot of copying. It would be possible to switch to using something similar to the Builder monad if need arises. However, since most protocols only have small bit fields, this should suffice for many cases.

Documentation

type BitPut = BitPutM ()

data BitPutM a

data BitPutT m a

runBitPut :: BitPut -> ByteString

runBitPutM :: BitPutM a -> (a, ByteString)

runBitPutT :: Monad m => BitPutT m a -> m (a, ByteString)

putBit :: Bool -> BitPut

Append a single bit

putBitT :: Monad m => Bool -> BitPutT m ()

putNBits :: (Integral a, Bits a) => Int -> a -> BitPut

Append the bottom n bits of the given bits value. In the case that more bits are requested than the value provides, this acts as if the value has as unlimited number of leading 0 bits.

putNBitsT :: (Monad m, Integral a, Bits a) => Int -> a -> BitPutT m ()

putBits :: (Integral a, Bits a) => a -> BitPut

Append a value. Note that this function is undefined for instances of Bits which have no fixed bitsize (like Integer)

putByteString :: ByteString -> BitPut

Append a ByteString

putLeftByteString :: (ByteString, Int) -> BitPut

Append a left aligned ByteString where ByteString has a partial byte with the given number of valid bits, from the MSB downwards. The number of such bits must be 0..7. (A normal ByteString, which all bytes full would use 0)