Data.Binary.Strict.ByteSet

Description

A ByteSet is a fast Set object for Word8's. The construction of these objects isn't terribly quick, but the member function should be about as good as you can get. Thus, you should use this when member is the most common operation

This object is designed to be imported qualified:

import qualified Data.Binary.Strict.ByteSet as BSet

Documentation

data ByteSet

Construction

empty :: ByteSet

An empty set

full :: ByteSet

The set contained all elements

singleton :: Word8 -> ByteSet

A set with a single element

fromList :: [Word8] -> ByteSet

range :: Word8 -> Word8 -> ByteSet

Construct a ByteSet containing all the elements from a to b, inclusive.

Combination

union :: ByteSet -> ByteSet -> ByteSet

intersection :: ByteSet -> ByteSet -> ByteSet

difference :: ByteSet -> ByteSet -> ByteSet

Uniary functions

complement :: ByteSet -> ByteSet

toList :: ByteSet -> [Word8]

member :: ByteSet -> Word8 -> Bool